# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenBooks API",
    "version": "5.0.0",
    "summary": "PotatoStack patched build of evan-buss/openbooks: IRC Highway eBook search and download, plus a REST API",
    "description": "PotatoStack v5 of evan-buss/openbooks (upstream base v4.5.0). Adds bearer-token authentication on every API route, a REST API (/api/v1) on top of the websocket-only upstream app, path-traversal hardening on all library endpoints, loopback-only bind by default, TLS verification for IRC, and vulnerability fixes in the dependency tree. All /api/v1 and legacy endpoints require a token when the server was started with OPENBOOKS_TOKEN (Authorization: Bearer header, X-OpenBooks-Token header, or ?token= query parameter). Only the static SPA and this document are unauthenticated."
  },
  "servers": [
    {
      "url": "/",
      "description": "The openbooks base path (e.g. http://host:8083/ or http://host:8083/openbooks/)"
    }
  ],
  "security": [
    {
      "openbooksToken": []
    }
  ],
...
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "summary": "Get current runtime settings",
        "description": "The download directory and persistence flag in effect for this running instance.",
        "operationId": "getSettings",
        "responses": {
          "200": {
            "description": "Settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update runtime settings",
        "description": "Change the download directory and/or persistence flag without a restart. The new directory must be absolute, existing or creatable, and writable; a rejected change leaves the settings untouched.",
        "operationId": "updateSettings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body or rejected download dir"
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "openbooksToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "The OPENBOOKS_TOKEN value. Accepted as Authorization: Bearer, X-OpenBooks-Token header, or ?token= query parameter. When the server has no token configured (single-user desktop mode) no credential is needed."
      }
    },
...
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "downloadDir": {
            "type": "string",
            "description": "Directory where downloads are saved; the library lives in downloadDir/books."
          },
          "persist": {
            "type": "boolean",
            "description": "Keep books after sending them to the client. When false, a served book is deleted afterwards."
          }
        },
        "required": [
          "downloadDir",
          "persist"
        ]
      },
      "SettingsUpdate": {
        "type": "object",
        "description": "Partial update: only the provided fields are changed.",
        "properties": {
          "downloadDir": {
            "type": "string"
          },
          "persist": {
            "type": "boolean"
          }
        }
      }
    }
  }
}